Fix -j/--json config setting
The -j/--json config option for specifying a config file did not work.
The problem was that after invoking getopt_long to parse the command
line options, the clerk_init() function overwrote the 'json' field of
the clrk_clerk_t object.
This change assigns the argument passed to the -j/--json option to a
temporary that is passed to clerk_init() and is used to initialize the
'json' string of the global clrk_clerk_t directly.