fix assertion failure when canceling user input for todo creation
When the user is about to create a new todo item but cancels the text
input an assertion failure is the result. The problem is that
clrk_input() can return NULL but clrk_todo_add() cannot handle this
case. This change fixes this issue.

Bonus: Use goto instead of return from this function to always trip over
       final LOG() statement.