fix todo loading error
Since the error code return check was added to check the result of
yajl_parse we see an error being reported everytime we load a file
containing todos.
The reason is a wrong buffer size (greater than the actual size) being
passed to the yajl_parse function which would then continue reading
uninitialized memory and trying to interpret it.
This change fixes the issue.