allow for cancelling of project creation
When creating a new project and cancelling the input (by means of
pressing ESC) a project with name "NONAME" is created. In my opinion
this is both unintuitive from a user's point of view (because the input
was cancelled) and inconsistent with respect to creation of TODOs where
aborting the input results in no new entry being created.
This change cancels project creation when the user decided to press ESC
when prompted for a project name. Also, since the only caller to
clrk_project_new no longer invokes it with a NULL pointer, it simplifies
the function a bit.