provide clrk_save handler for SIGUSR1 signal
When a clerk instance is running but the user has no access to it (e.g.,
because he/she is logged in remotely and clerk is running in a different
session), it might still be desireable to terminate this instance and
start a new one. However, until now the user had no chance to save any
changes he/she made in clerk beforehand.
This change introduces a signal handler for SIGUSR1 that can be used to
save the changes that were made previously.