time stamp log output
Clerk's logging feature helps when debugging certain issues. However, to
ease investigation of certain issues it might be helpful to have a time
stamp associated to each line of the log.
This change prepends a time stamp of the form 'hh:mm:ss' to each log
message (where 'hh' represents the current hour of the day, 'mm' the
current minute, and 'ss' the current second).