add SIGTERM and SIGINT support
When clerk is killed by SIGTERM or SIGINT it performs no cleanup and exits the
hard way. State such as the cursor visibility or the current console colors
are not reset to their original values.
This change adds a handler for SIGTERM and SIGINT. Arguably, a handler for
SIGINT is not really required since Ctrl+C (the usual way to receive a SIGINT)
is already taken care of by termbox. However, since there might be other ways
to send this signal to clerk an exit handler is registered for it as well.